🕸️ Weaver: Refactor SkillsEditor using Custom Hook Pattern
